Added OneDrive setup running check

+++ scripts/4.remove-onedrive.ps1
@@ -22,6 +22,13 @@ New-PSDrive HKU Registry HKEY_USERS | Out-Null
 
 #=============================================================================
 
+$oneDriveSetup = Get-Process "OneDriveSetup" -ErrorAction SilentlyContinue
+if ($oneDriveSetup) {
+  Write-Host "OneDriveSetup is still running - uninstall skipped" -BackgroundColor DarkYellow
+  $skipOneDriveUnintall=$true
+  return
+}
+
 Stop-Process -name OneDrive
 
 # Remove OneDrive
